Version 7.3.108 (16.09.2023) - UPDATED: Dependencies - FIXED: Display bug in breaking an alliance (multiple alliances are shown as broken, although only one changed)- - FIXED: Map planetinfopanel needs to mouse transparent if invisible - FIXED: If a jumpship attacks, the icon is moved back to the starting system on save... the jumpship level icon needs to be moved as well - CHANGED: Delete all unneeded columns in table 'user'

+++ net.clanwolf.starmap.client/src/main/java/net/clanwolf/starmap/client/util/Tools.java
@@ -471,9 +471,13 @@ public final class Tools {
 //			System.out.println("diff: " + diff);
 
 			if (numberOfDays * 24 * 60 * 60 * 1000 < diff) {
-				boolean success = file.delete();
-				if (!success) {
-					logger.info("File could not be deleted: " + file.getAbsolutePath());
+				if (!file.getName().endsWith(".mp4")) {
+					boolean success = file.delete();
+					if (!success) {
+						logger.info("File could not be deleted: " + file.getAbsolutePath());
+					}
+				} else {
+					logger.info("Video files are not deleted in cache to save traffic. To empty the cache use purge!");
 				}
 			} else {
 				// file will be left
